refactor: Voucher Interface를 abstract class로 변경
