+++ src/main/java/com/example/voucher_manager/domain/voucher/FixedAmountVoucher.java
@@ -2,34 +2,25 @@ package com.example.voucher_manager.domain.voucher;
 
 import java.util.UUID;
 
-public class FixedAmountVoucher implements Voucher {
-    private final UUID voucherId;
+public class FixedAmountVoucher extends Voucher {
     private final Long discountPrice;
     private final VoucherType voucherType;
-    private UUID ownerId;
 
     public FixedAmountVoucher(UUID voucherId, Long discountPrice, VoucherType voucherType) {
-        this.voucherId = voucherId;
+        super(voucherId, null);
         this.discountPrice = discountPrice;
         this.voucherType = voucherType;
-        this.ownerId = null; // 주인 미배정 상태
     }
 
     public FixedAmountVoucher(UUID voucherId, Long discountPrice, VoucherType voucherType, UUID ownerId) {
-        this.voucherId = voucherId;
+        super(voucherId, ownerId);
         this.discountPrice = discountPrice;
         this.voucherType = voucherType;
-        this.ownerId = ownerId;
     }
 
     @Override
     public String toString() {
         return "[VoucherType : FixedAmountVoucher," + " VoucherId : " + voucherId + ", DiscountPrice : " + discountPrice + "$]";
-    }
-
-    @Override
-    public UUID getVoucherId() {
-        return voucherId;
     }
 
     @Override
@@ -52,8 +43,4 @@ public class FixedAmountVoucher implements Voucher {
         return discountPrice;
     }
 
-    @Override
-    public UUID getOwnerId() {
-        return ownerId;
-    }
 }
